Bluetooth: Fix type of max latency and min timeout used in NVRAM
This patch fixes regression caused by commit
b6264c2464da28359291eb6785479b236bfc11e7 where Max latency and Min timeout
were being stored as 8 bit unsigned integers, instead of 16 bits unsigned
integers, resulting in loss of data. Max Latenency and Min timeout values
were not being copied properly resulting in wrong values being sent during
BTA_HH_ADD_DEV_EVT. Also some other side effects left from the commit that
entry was not being removed from NVRAM and duplicate entry additions have
also been fixed in this patch.

CRs-Fixed: 533988

+++ btif/include/btif_storage.h
@@ -236,8 +236,8 @@ bt_status_t btif_storage_add_hid_device_info(bt_bdaddr_t *remote_bd_addr,
                                                     UINT16 attr_mask, UINT8 sub_class,
                                                     UINT8 app_id, UINT16 vendor_id,
                                                     UINT16 product_id, UINT16 version,
-                                                    UINT8 ctry_code, UINT8 ssr_max_lat,
-                                                    UINT8 ssr_min_tout, UINT16 dl_len,
+                                                    UINT8 ctry_code, UINT16 ssr_max_lat,
+                                                    UINT16 ssr_min_tout, UINT16 dl_len,
                                                     UINT8 *dsc_list);
 
 /*******************************************************************************
